requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
Django==1.11.21
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djproxy==2.3.4
elasticsearch==7.0.2
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
whitenoise==4.1.2

